fix make parallel builds
The makefiles used recursive make invocation by calling `make`.
This is bad practise, as the calling make cannot be sure that the
process invoked by calling `make` is actually a make. It therefore
cannot pass the options required to get the job server to work to make,
and as a result builds are limited to a single thread, which is slow.
Luckily, this bug is easily fixed by replacing all occurrences of `make`
with `$(MAKE)`.